fix: use correct validation error for missing required fields
